requirements.txt
streamlit
requests
openai
sentence-transformers
numpy
faiss-cpu
pymupdf
langchain
langchain-community
torch
